+++ CHANGELOG.md
@@ -14,7 +14,38 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 -
 
 ### Added
-
+- Allowed setting playback `ex.Sound.duration` which will limit the amount of time that a clip plays from the current playback position.
+- Added a new lightweight `ex.StateMachine` type for building finite state machines
+  ```typescript
+  const machine = ex.StateMachine.create({
+    start: 'STOPPED',
+    states: {
+      PLAYING: {
+        onEnter: () => {
+          console.log("playing");
+        },
+        transitions: ['STOPPED', 'PAUSED']
+      },
+      STOPPED: {
+        onEnter: () => {
+          console.log("stopped");
+        },
+        transitions: ['PLAYING', 'SEEK']
+      },
+      SEEK: {
+        transitions: ['*']
+      },
+      PAUSED: {
+        onEnter: () => {
+          console.log("paused")
+        },
+        transitions: ['PLAYING', 'STOPPED']
+      }
+    }
+  });
+  ```
+- Added `ex.Sound.seek(positionInSeconds)` which will allow you to see to a place in the sound, this will implicitly pause the sound
+- Added `ex.Sound.getTotalPlaybackDuration()` which will return the total time in the sound in seconds.
 - Allow tinting of `ex.Sprite`'s by setting a new `tint` property, renderers must support the tint property in order to function.
   ```typescript
   const imageSource = new ex.ImageSource('./path/to/image.png');
@@ -39,7 +70,13 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 
--
+- Fixed issue with `ex.ArcadeSolver` based collisions where colliders were catching on seams when sliding along a floor of multiple colliders. This was by sorting contacts by distance between bodies.
+  ![sorted-collisions](https://user-images.githubusercontent.com/612071/172401390-9e9c3490-3566-47bf-b258-6a7da86a3464.gif)
+
+- Fixed issue with `ex.ArcadeSolver` where corner contacts would zero out velocity even if the bodies were already moving away from the contact "divergent contacts".
+  ![cancel-velocity-fix](https://user-images.githubusercontent.com/612071/172500318-539f3a36-31ae-4efc-b6ab-c4524b297adb.gif)
+
+- Fixed issue where `ex.Sound` wasn't being paused when the browser window lost focus
 
 ### Updates
 
@@ -190,7 +227,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Added new display mode `ex.DisplayMode.FitScreenAndFill`.
 - Added new display mode `ex.DisplayMode.FitContainerAndZoom`.
 - Added new display mode `ex.DisplayMode.FitScreenAndZoom`.
-
+- Added the ability to select variable duration into Timer constructor. 
 ### Fixed
 
 - Fixed unreleased bug where CompositeCollider components would not collide appropriately because contacts did not have unique ids